llm-saas-backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# HTTP Client (for OpenAI - optional)
httpx==0.27.2
openai==1.51.2

# Environment
python-dotenv==1.0.1

# Logging
structlog==24.4.0

# Testing (dev)
pytest==8.3.3
pytest-asyncio==0.24.0